scripts/requirements.txt
certifi>=2024.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-chess>=1.999
sqlalchemy>=2.0.0
zstandard>=0.21.0
